[update]: logo 타입스크립트 마이그레이션
